# docs/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cloacina-server API",
    "description": "REST API for the Cloacina workflow-orchestration server. WebSocket protocols (execution events, accumulator ingest, reactor commands, substrate delivery) are specified in the WebSocket Protocol reference of the documentation site.",
    "contact": {
      "name": "Dylan Storey",
      "email": "dylan.storey@gmail.com"
    },
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.7.0"
  },
  "paths": {
...
    "/v1/auth/keys": {
      "get": {
        "tags": [
          "keys"
        ],
        "summary": "GET /auth/keys — list all API keys (no hashes or plaintext).\nRequires admin role.",
        "operationId": "list_keys",
        "responses": {
          "200": {
            "description": "All keys (no hashes or plaintext)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_KeyInfo"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      },
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "POST /auth/keys — create a new API key.",
        "description": "Requires admin role. Non-admin keys cannot create keys with higher\npermissions than their own (prevents privilege escalation).\nReturns the plaintext key exactly once. It cannot be retrieved again.",
        "operationId": "create_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Key created — plaintext returned exactly once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyCreatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/auth/keys/{key_id}": {
      "delete": {
        "tags": [
          "keys"
        ],
        "summary": "DELETE /auth/keys/:key_id — revoke an API key.\nRequires admin role within tenant (or god mode).",
        "operationId": "revoke_key",
        "parameters": [
          {
            "name": "key_id",
            "in": "path",
            "description": "Key UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyRevokedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid key ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Key not found or already revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/v1/auth/ws-ticket": {
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "POST /auth/ws-ticket — exchange a Bearer token for a single-use WebSocket ticket.",
        "description": "Returns a short-lived ticket that can be used as a query parameter for\nWebSocket upgrade requests, avoiding long-lived API keys in URLs.",
        "operationId": "create_ws_ticket",
        "responses": {
          "200": {
            "description": "Single-use WebSocket ticket",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WsTicketResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
    "/v1/tenants/{tenant_id}/keys": {
      "post": {
        "tags": [
          "keys"
        ],
        "summary": "POST /tenants/:tenant_id/keys — create a key scoped to a tenant.\nAdmin-only: only is_admin keys can create tenant-scoped keys.",
        "operationId": "create_tenant_key",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "description": "Tenant identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Tenant-scoped key created — plaintext returned exactly once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyCreatedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Admin (god-mode) key required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateKeyRequest": {
        "type": "object",
        "description": "Request body for `POST /auth/keys` and `POST /tenants/{tenant_id}/keys`.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/KeyRole"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Standardized error response body. Every non-2xx response from\n`cloacina-server` carries this shape; the request correlation ID is in\nthe `x-request-id` response header, not the body.",
        "required": [
          "error",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code (e.g. `invalid_pagination`,\n`workflow_not_found`)."
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "KeyCreatedResponse": {
        "type": "object",
        "description": "`201 Created` body for a new API key. The plaintext `key` is returned\nexactly once — it cannot be retrieved again.",
        "required": [
          "id",
          "name",
          "key",
          "permissions",
          "is_admin",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "description": "RFC 3339 timestamp."
          },
          "id": {
            "type": "string",
            "description": "Key UUID."
          },
          "is_admin": {
            "type": "boolean",
            "description": "God-mode flag — never granted via the create endpoints."
          },
          "key": {
            "type": "string",
            "description": "One-time plaintext API key."
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "string",
            "description": "Role string: `read` | `write` | `admin`."
          },
          "tenant_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Tenant scope; `null` for global keys."
          }
        }
      },
...
      "KeyRevokedResponse": {
        "type": "object",
        "description": "`DELETE /auth/keys/{key_id}` response.",
        "required": [
          "status",
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "UUID of the revoked key."
          },
          "status": {
            "type": "string",
            "description": "Always `\"revoked\"`."
          }
        }
      },
      "KeyRole": {
        "type": "string",
        "description": "Allowed roles for API keys.",
        "enum": [
          "admin",
          "write",
          "read"
        ]
      },
...
      "ListResponse_KeyInfo": {
        "type": "object",
        "description": "Unified list envelope (CLOACI-T-0594 / API-03): every list endpoint\nreturns `{items, total}`. `total` is best-effort — it equals the\nreturned page size when the server doesn't run a separate COUNT.",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One row in the key list (`GET /auth/keys`). No hashes or plaintext.",
              "required": [
                "id",
                "name",
                "permissions",
                "is_admin",
                "created_at",
                "revoked"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "description": "RFC 3339 timestamp."
                },
                "id": {
                  "type": "string",
                  "description": "Key UUID."
                },
                "is_admin": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "permissions": {
                  "type": "string",
                  "description": "Role string: `read` | `write` | `admin`."
                },
                "revoked": {
                  "type": "boolean"
                },
                "tenant_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Tenant scope; `null` for global keys."
                }
              }
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
      "WsTicketResponse": {
        "type": "object",
        "description": "`POST /auth/ws-ticket` response — a single-use, short-lived ticket for\nWebSocket upgrade auth (avoids long-lived API keys in URLs).",
        "required": [
          "ticket",
          "expires_in_seconds"
        ],
        "properties": {
          "expires_in_seconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "ticket": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "api_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Cloacina API key, sent as `Authorization: Bearer <key>`. Tenant scope is carried by the key itself; tenant-scoped routes additionally name the tenant in the path."
      }
    }
  },
...
}